requirements.txt
# ── 핵심 의존성 (필수) ───────────────────────────────────────────────
moviepy>=1.0.3
opencv-python>=4.8.0
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.0.0

# ── AI / API ─────────────────────────────────────────────────────────
# Qwen VLM 하이라이트 분석 (선택, DASHSCOPE_API_KEY 환경변수 필요)
aiohttp>=3.9.0

# ── 자세 분석 (선택) ──────────────────────────────────────────────────
# MediaPipe 베스트컷 프레임 선택 (없으면 첫 번째 후보 사용)
# mediapipe>=0.10.0

# ── 화질 개선 (선택, GPU 권장) ───────────────────────────────────────
# macOS Apple Silicon (M1/M2/M3) MPS 가속:
# pip install torch torchvision
#
# NVIDIA GPU (Linux/Windows) CUDA 가속:
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
#
# Real-ESRGAN 4x 업스케일링 모델 로더:
# pip install spandrel>=0.4.0
#
# 모델 파일 다운로드 (처음 1회, 약 64MB):
# mkdir -p models
# curl -L -o models/RealESRGAN_x4plus.pth \
#   https://github.com/xinntao/Real-ESRGAN/releases/download/v0.1.0/RealESRGAN_x4plus.pth
